requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
av==12.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
ctransformers==0.2.27
ctranslate2==4.3.1
distro==1.9.0
edge-tts==6.1.12
faster-whisper==1.0.2
filelock==3.15.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
keyboard==0.13.5
mpmath==1.3.0
multidict==6.0.5
numpy==2.0.0
nvidia-cublas-cu12==12.5.2.13
nvidia-cuda-nvrtc-cu12==12.5.40
nvidia-cuda-runtime-cu12==12.5.39
nvidia-cudnn-cu12==9.1.1.17
onnxruntime==1.18.0
openai==1.34.0
packaging==24.1
protobuf==5.27.1
py-cpuinfo==9.0.0
pyaes==1.6.1
PyAudio==0.2.14
pydantic==2.7.4
pydantic_core==2.18.4
pygame==2.5.2
pyreadline3==3.4.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.19.2
python-dotenv==1.0.1
python-telegram-bot==21.3
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
sniffio==1.3.1
sympy==1.12.1
tokenizers==0.19.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
wheel==0.43.0
yarl==1.9.4
